Fix Mac build.
R=ager@google.com
BUG=

Review URL: https://codereview.chromium.org/894383002

+++ tests/service_tests/person/cc/struct.cc
@@ -98,7 +98,7 @@ int64_t Builder::InvokeMethod(ServiceId service, MethodId method) {
   // Mark the request as being segmented.
   *reinterpret_cast<int64_t*>(buffer + 32) = 1;
   ServiceApiInvoke(service, method, buffer, size);
-  int result = *reinterpret_cast<int64_t*>(buffer + 32);
+  int64_t result = *reinterpret_cast<int64_t*>(buffer + 32);
   free(buffer);
   return result;
 }
